+++ src/game/BossSpider.cpp
@@ -27,15 +27,15 @@
 
 BossSpider::BossSpider()
 :Enemy()
-{	
+{
 	setImage("spider.png");
 	walk_speed = 100.0f;
-	
+
 	speed_x = 0;
 	speed_y = 0;
 	dying = false;
 	life = 5;
-	
+
 	width = 60;
 	height = 64;
 	xOrigin = width/2;
@@ -45,12 +45,12 @@ BossSpider::BossSpider()
 	lastShot = 0;
 	shootInterval = 0.5f;
 	time = 0;
-	
+
 	patrolInterval = 1.5f;
 	patrolTime = 0;
-	
+
 	Animation * tmp;
-	
+
 	//pick a random death sound
 	int sound_num = rand() % 19;
 	sound_num += 1;
@@ -58,12 +58,12 @@ BossSpider::BossSpider()
 	std::stringstream out;
 	out << sound_num;
 	s = out.str();
-	
+
 	std::string sound_file = s + "-BugSplat.ogg";
 	//cout << sound_file;
-	fireSound = soundCache[sound_file];	
-	
-	
+	fireSound = soundCache[sound_file];
+
+
 	tmp = addAnimation("walk");
 	tmp->addFrame(0, .2f);
 	tmp->addFrame(1, .2f);
@@ -71,13 +71,13 @@ BossSpider::BossSpider()
 	tmp->addFrame(4, .2f);
 	tmp->addFrame(5, .2f);
 	tmp->setDoLoop(true);
-	
+
 	tmp = addAnimation("die");
 	tmp->addFrame(0, .07f);
-	
+
 	tmp = addAnimation("hurt");
 	tmp->addFrame(0, 0.07f);
-	
+
 	setCurrentAnimation("walk");
 }
 
@@ -85,15 +85,15 @@ void BossSpider::update(float dt) {
 	if(!dying) {
 		time += dt;
 		patrolTime += dt;
-		
+
 		//setCurrentAnimation("walk");
 		const int speed_gravity = 960;
 		const float vision_range = 320;
 		const float vision_min_range = 32;
-		
+
 		speed_y += speed_gravity*dt;
 		if(isGrounded()) speed_y = 0;
-		
+
 		if(facing_direction == FACING_LEFT) {
 			speed_x = -walk_speed * dt;
 			if(patrolTime > patrolInterval) {
@@ -109,15 +109,15 @@ void BossSpider::update(float dt) {
 		}
 
 		move(speed_x, speed_y);
-		
+
 		updateSpriteFacing();
-		
+
 		if(lastShot + shootInterval < time) {
 			lastShot = time;
-			EnemyCentipedeProjectile * projectile = 
-				new EnemyCentipedeProjectile(facing_direction, pos_x, pos_y - 20);
+			EnemyCentipedeProjectile * projectile =
+				new EnemyCentipedeProjectile(facing_direction, (int)pos_x, int(pos_y - 20.0f));
 		}
-		
+
 		checkCollisions();
 	}
 }
@@ -129,11 +129,11 @@ void BossSpider::draw() {
 
 void BossSpider::doDamage(float damage) {
 	life -= damage;
-	if(life <= 0) 
+	if(life <= 0)
 		die();
-	else 
+	else
 		setCurrentAnimation("hurt");
-	
+
 }
 
 void BossSpider::die() {
@@ -152,7 +152,7 @@ void BossSpider::onAnimationComplete(std::string anim) {
 		CollectibleEnergyBall * ball = new CollectibleEnergyBall();
 		ball->setPos(pos_x-16, pos_y-16);
 	}
-	
+
 	if(anim == "hurt") {
 		setCurrentAnimation("walk");
 	}
